First test. Still broken.

+++ src/core/components/Sprite.cc
@@ -17,15 +17,15 @@ void Sprite::draw(s2x::Renderer& renderer, const Transform& transform)
     auto size = this->size();
 
     SDL_Rect r = { 
-        coords.position.x / coords.scaleX, 
-        coords.position.y / coords.scaleY, 
-        size.width, 
-        size.height 
+        (int)(coords.position.x / coords.scaleX), 
+        (int)(coords.position.y / coords.scaleY), 
+        (int)(size.width), 
+        (int)(size.height)
     };
 
     SDL_Point ra = { 
-        coords.rotationAnchor.x, 
-        coords.rotationAnchor.y 
+        (int)coords.rotationAnchor.x, 
+        (int)coords.rotationAnchor.y 
     };
 
     renderer.copy(texture, r, coords.rotation, ra, coords.scaleX, coords.scaleY);
